// src/types/menus/IMenu.ts
export enum EMenuCategoryType {
  MAIN_ITEMS = 'MAIN_ITEMS',
  PIZZA = 'PIZZA',
}

export enum EPortionSize {
  SERVES_1 = 'SERVES_1',
  SERVES_2 = 'SERVES_2',
  SERVES_3 = 'SERVES_3',
  SERVES_4 = 'SERVES_4',
  NOT_APPLICABLE = 'NOT_APPLICABLE',
}

export enum EDietaryRestriction {
  GLUTEN_FREE = 'GLUTEN_FREE',
  LAC_FREE = 'LAC_FREE',
  ORGANIC = 'ORGANIC',
  SUGAR_FREE = 'SUGAR_FREE',
  VEGAN = 'VEGAN',
  VEGETARIAN = 'VEGETARIAN',
}

export enum EDishClassification {
  ALCOHOLIC_DRINK = 'ALCOHOLIC_DRINK',
  FROSTY = 'FROSTY',
}

export enum EImportMenuStatus {
  NOT_IMPORTED = 'NOT_IMPORTED',
  SCHEDULED = 'SCHEDULED',
  PROCESSING = 'PROCESSING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}